#60 refactor: 분석 결과 조회 StudyRecord -> Study 이동
