+++ core/data/src/main/java/com/ds/studify/core/data/datasource/StudyDataSource.kt
@@ -2,6 +2,7 @@ package com.ds.studify.core.data.datasource
 
 import com.ds.studify.core.data.dto.request.RequestRecordDto
 import com.ds.studify.core.data.dto.response.BaseResponse
+import com.ds.studify.core.data.dto.response.ResponseAnalysisDto
 import com.ds.studify.core.data.service.StudyService
 import kotlinx.serialization.json.JsonElement
 import javax.inject.Inject
@@ -11,4 +12,7 @@ class StudyDataSource @Inject constructor(
 ) {
     suspend fun postRecord(request: RequestRecordDto): BaseResponse<JsonElement?> =
         studyService.postRecord(request = request)
+
+    suspend fun getAnalysis(studyRecordId: Int): BaseResponse<ResponseAnalysisDto> =
+        studyService.getAnalysis(studyRecordId)
 }
