+++ core/data/src/main/java/com/ds/studify/core/data/datasource/StudyRecordDataSource.kt
@@ -4,7 +4,6 @@ import com.ds.studify.core.data.dto.response.BaseResponse
 import com.ds.studify.core.data.dto.response.ResponseCalendarDailyDto
 import com.ds.studify.core.data.dto.response.ResponseCalendarMonthlyDto
 import com.ds.studify.core.data.dto.response.ResponseFeedbackDto
-import com.ds.studify.core.data.dto.response.ResponseAnalysisDto
 import com.ds.studify.core.data.dto.response.ResponseHomeDto
 import com.ds.studify.core.data.dto.response.ResponsePieChartDto
 import com.ds.studify.core.data.service.StudyRecordService
@@ -27,9 +26,4 @@ class StudyRecordDataSource @Inject constructor(
 
     suspend fun getPieChart(studyRecordId: Long, tab: String): BaseResponse<List<ResponsePieChartDto>> =
         studyRecordService.getPieChart(studyRecordId = studyRecordId, tab = tab)
-        
-    suspend fun getAnalysis(
-        studyRecordId: Int
-    ): BaseResponse<ResponseAnalysisDto> =
-        studyRecordService.getAnalysis(studyRecordId)
 }
