+++ core/data/src/main/java/com/ds/studify/core/data/repository/StudyRecordRepository.kt
@@ -3,7 +3,6 @@ package com.ds.studify.core.data.repository
 import com.ds.studify.core.domain.entity.CalendarDailyEntity
 import com.ds.studify.core.domain.entity.CalendarMonthlyEntity
 import com.ds.studify.core.domain.entity.FeedbackEntity
-import com.ds.studify.core.domain.entity.AnalysisEntity
 import com.ds.studify.core.domain.entity.HomeEntity
 import com.ds.studify.core.domain.entity.PieChartEntity
 
@@ -17,6 +16,4 @@ interface StudyRecordRepository {
     suspend fun getFeedback(studyRecordId: Long): Result<FeedbackEntity>
 
     suspend fun getPieChart(studyRecordId: Long, tab: String): Result<List<PieChartEntity>>
-  
-    suspend fun getAnalysis(studyRecordId: Int): Result<AnalysisEntity>
 }
